Complete overhaul of API to clear up node vs key vs value confusion
There is no more "value" object - values are strictly properties
of their owning keys. Everything is accessed via the parent key,
and all keys are formally called "subkeys" in the API so that no
one is confused between the key (as in, the thing in the folder in
regedit) and the "key" (as in a value is a <key, value> pair).